requirements.txt
avro==1.8.1
cwltool==1.0.20160811184335
html5lib==0.999999999
isodate==0.5.4
keepalive==0.5
mistune==0.7.3
pyparsing==2.1.8
PyYAML==3.11
rdflib==4.2.1
rdflib-jsonld==0.4.0
requests==2.11.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.11.11
schema-salad==1.16.20160810195039
shellescape==3.4.1
six==1.10.0
SPARQLWrapper==1.7.6
typing==3.5.2.2
webencodings==0.5
